requirements.txt
tqdm==4.47.0
openpyxl==3.0.4
python-dateutil==2.8.1
mysql-connector-python==8.0.15
pyodbc==4.0.30
XlsxWriter==1.3.7